split the AdminLayout component into smaller, manageable components
