+++ pages/admin/department.tsx
@@ -1,4 +1,4 @@
-import AdminLayout from '@components/Layout/AdminLayout';
+import AdminLayout from '@components/layout/admin';
 import DepartmentView from '@components/views/admin/department';
 
 export default () => (
