+++ pages/admin/users.tsx
@@ -1,4 +1,4 @@
-import AdminLayout from '@components/Layout/AdminLayout';
+import AdminLayout from '@components/layout/admin';
 import UserView from '@components/views/admin/users';
 
 export default () => (
